src: move TabContainer file to Tabs folder
