+++ app/containers/BuyPage/index.js
@@ -12,9 +12,9 @@ import MDCAppBar from '../../components/AppBar';
 import MDCHeader from '../../components/AppBar/Header';
 import MDCTabBar from '../../components/AppBar/TabBar';
 import ErrorBoundary from '../../components/ErrorBoundary';
+import { TabContainer } from '../../components/Tabs';
 import { NavigationLayout } from '../Layout';
 import HeaderTabs from './components/HeaderTabs';
-import TabContainer from './components/TabContainer';
 import TestSwap from './components/TestSwap';
 import MyOrders from './MyOrders';
 import PlaceOrder from './PlaceOrder';
